Collaboration feature
Enables Google Docs-like collaboration between TIC-80 users.

Adds a new console command COLLAB, which connects to a remote
collaboration server, given a URL that contains an alphanumeric session
ID.

When connected, parts of the cart that differ from the server are
highlighted in yellow, and PUSH and PULL toolbar buttons allow copying
the state from the client to the server or vice versa.

The collaboration server is also included in this commit. It is a simple
Python HTTP server which maintains each collaborated cart as a file.