Map view fixes when wrapping around the edge
- When trying to select across the edge of the map, the selection
rectangle no longer jumps and stretches across the entire map.

The mouse coordinates were being normalized prior to expanding the
selection extent, so for example a 1,1 start point and a -1,-1 end point
would evaluate to 1,1..W-1,H-1, turning a 2x2 selection into the entire
map!.

It's actually fine for the selection rectangle to go outside the valid
range because the copy/paste functions all normalize the individual
gets/sets.

- Don't allow scrolling into negative offsets in the world view. This
breaks the grid drawing code temporarily when you switch back to the
map.

- When the map rectangle wraps around in the world view, draw all the
wrapped quadrants. This makes it clearer how the map wraps around and
avoids "losing" the viewport when you scroll slightly left or up.