Code cursor undo change
Fixes the cursor getting out of sync when undoing and redoing in the
code window.

Under the hood this is quite complicated, but I hope it produces the
desired effect and also it may enable some other undo improvements in
the future.

The principle is that when undoing a code change, you want the cursor to
end up where it was before the code change, and when redoing a code
change, you want the cursor to end up where it was _after_ the change.

To accomplish this, cursor undo is stored with a TAG (before or after).
When the code is modified by any means, both a BEFORE and an AFTER
cursor tag are added. When undoing, the cursor history is rewound to the
next BEFORE tag, and when redoing, it is forwarded to the next AFTER
tag.

One issue that is handled is the case of an edit after undoing. Normally,
edits creates a BABA... alternation in the cursor undo history.
If the user makes some changes, undoes a change, and then makes more
changes, you can get a BABBA where the BB is the point of the undone
change. To resolve this, when the tag being added to the history matches
the latest tag in the history, the latest tag is replaced so there won't
be two of the same tag in the sequence. I'm on the fence about whether
this behavior belongs in the history system or in the code, but there
are no other users of the tags yet so I'll leave it as written. In the
future the tags system might be used to implement "word" undo ala
VSCode, or stroke undo (instead of per-pixel) in the sprite editor. But
I'm not sure, these features might better benefit from an overhaul of
the undo system, I'd need to research other possibilities.

As a simplification, code changes are now flagged in a bool, which
triggers all the syntax highlighting, undo, etc. This may have fixed a
few places where these were applied redundantly.