Fix Android black screen when paused
When switching away from TIC-80, the OpenGL context is destroyed by the
system and needs to be recreated along with all resources (textures and
shaders).

SDL2 makes an attempt to save and restore the active GL context when
pausing and resuming the activity, but it fails on my device, and the
Android and SDL2 documentation indicate that it's a best-effort feature.

Because the keyboard and gamepad images are lost when the textures are
destroyed, backing buffers are allocated to store them when they are
first filled out.

Additionally, SDL_gpu needs to be fully destroyed and recreated to
resolve the problem on my Pixel 3a XL. Just recreating the textures and
shader and calling GPU_ResetRenderState is not enough.

(Ideally, this entire problem is something that would be handled by
the SDL_gpu library, but it has no provisions for it)