Update Fennel keywords to include new 0.4.x additions.
Also add fennelfriend compiler helpers into options during
compilation.

This gives us better error messages, but the "pinpointing" of the
error which we have in CLI fennel does not work here, because it
depends upon being able to io.open the source file to find the source
code which is throwing the error.

We need some way of pinpointing errors when the source is only
available as a string.