Activate strict mode in Fennel when strict: true header is found.
Global mode makes it so that it is a compile error to refer to a
global which doesn't exist yet. Using (global foo {}) will declare foo
as a global which can be used after the declaration, or you can always
bypass the checks using _G.foo.

This will catch a lot of errors earlier on during the compilation
process instead of at runtime. This should have been on by default
from the start. But better late than never. =)

We change the demo Fennel cart to add the strict header by default for
new games, but it's easy to delete for users who want the more
Lua-like global behavior.