Create SECURITY.md
I went to see what our security policy was, and we didn't have one, so this is a beginning.  I just used the default template and filled in the very basics.

Not sure if you'd prefer a nicer email like `security@tic80.com`?  

Also is there any reason someone should still be running 0.80 or 0.70?  Could we mark anything below 0.90 as obsolete?  This also has wiki implications as we don't need to mention "added in version 0.2" if someone isn't using version 0.1 anymore, etc...