Fix segfault in player-sdl
There is a bug in that tic80_tick doesnt set up the counter and freq callbacks
properly which leads to a segfault.

This is probably not a good solution and it only works for player-sdl and the
same issue is probably also in libretro which needs a proper fix too.