Bypass checks which disable Fennel friendly compiler errors.
In Fennel 1.3.0, it checks to see whether _G.io.read exists before
proceeding with friendly compiler error messages, because it used to
have to read the source from disk in order to provide these
messages. That is no longer needed; it can use the source provided as
a string, but the check was accidentally left in place.

This change bypasses that check with a fake io table in order to get
improved compiler error message and parse error reporting.

It should be removed once Fennel 1.3.1 or 1.4.0 is brought in.