JS API: Fix spr colorkey default parameter
In case that no parameter or undefined is passed to the spr function in
JS, the parameter would evaluate to 0 due to a wrong default value in
the code path when colorkey is not an array. In case of an array the
default value would be -1, which is also specified in the wiki.

We now pass -1 in case the parameter is not an array.

See: #2291