Python API: Fix float cast for all functions
The current implementation does not implicitly cast integer values into
float values if passed to a method which expects float values in the
python API. From a user perspective it would make sense that "func(1)"
and "func(1.0)" would be accepted.

To fix the problem a "to_float_wrapper" function is implemented that
will first check if the parameter passed is an integer and if yes
implicitly convert the integer into a double.

See: #2296