SDL: don't ignore briefly pressed keys
Some keyboards send KEY_DOWN/KEY_UP immediately in
sequence to support things like dual-purpose keys.
This change updates the SDL key handler to detect
when a key is pressed and released within a single
frame, and to propagate that key's state as down
for the duration of the frame, rather than just
drop the event entirely.