Implement "resume reload" command.
When resuming, you can pass "reload" as the argument, which causes the
cart's code to be reloaded before resuming the game.

This allows for a much smoother iterative development style where you
can pause the game, make changes, and see the effect of those changes
without restarting the game. It does require some special handling in
the cart to support saving state to a global and avoiding that state
being reinitialized if it's already set; for instance:

    state = state or {x=25, y=100, shots={}}

This implements it in a way that will be supported by any script that
already supports eval.