Libretro cursorfix (#2822)
* [libretro] draw cursor directly to screen

This fixes some bugs where the cursor were not drawn in some games, and
also causing a cursor trail in other games.

* [libretro] Makes the cursor to start hidden

As many games don't use mouse, we keep it hidden while it was not moved.
Those wanting to take screenshots of the titlescreen don't risk taking a
screenshot of the cursor.