[Android] Update gradle and gradlew to v9.0.0 (#2836)
* [Android] Update gradle and gradlew to v9.0.0

As we update gradle-wrapper.jar, it is desirable to confirm the
authenticity.

It can be checked by verifying:
sha256sum gradle/wrapper/gradle-wrapper.jar

to be identical to the one available at:
http://services.gradle.org/distributions/gradle-9.0.0-wrapper.jar.sha256

that should be:

76805e32c009c0cf0dd5d206bddc9fb22ea42e84db904b764f3047de095493f3

* [Android] updates JDK to v17

* [Android] Updates NDK to version r23

* [Android] informs NDK version in gradle file

The newer versions of Gradle requires us to explicitly inform the NDK
version we use.

* readme and comments